Remove public
We don’t want to track the public folder because it’s automatically built every time that we make a change in our website (e.g., change the template or add a post). This pollutes our commits by making it harder to see the actual changes we made to our templates/posts/other parts of the site.