The menu changes depending on the current page
We can’t put all the package folders under the R folder just yet because Hugo can’t figure out which package is which when they’re grouped under a common folder. If they’re all under a common folder, Hugo assumes each package’s section is just “R.” Instead, we want the section of each package to be the name of the package (e.g., the section of randomizr will be “randomizr”).