Fix of bad, bad, bad Travis bug
It is very, very, very important to keep all folder names lowercase. Uppercase filenames create bad bugs when building on Travis. When Hugo runs on Travis and there are uppercase folder names, Hugo will create two directories in the public directory for each folder in the content directory: one folder has its original name, and the other folder has the same name but in lowercase. This is bad because we do not want these weird duplicate folders.